requirements.txt
streamlit
pandas
scikit-learn
xgboostgit init
